fix "Potential null pointer access" issues
After individually checking all “Potential null pointer access” it turns out most of them are seem to be false messages (that’s why it is called “Potential…”). They exist are in “original” code and were not introduced by SwingJS changes. 

Therefore I changed the “Potential null pointer access” from error to warning and used to “@SuppressWarning” to hide the now verified message.

If we later write our own code and introduce a “null pointer” issue these warnings will still be active and warn us. Then we either fix the bug or use @SuppressWarning if we verify this is a false message.

# Other Errors

All errors are fixed.

# Imports

I started to change the “import xxx.*” to single imports.